Create AbstractRefreshToken - mapped as MappedSuperClass
