+++ Entity/RefreshToken.php
@@ -23,7 +23,7 @@ use Symfony\Bridge\Doctrine\Validator\Constraints\UniqueEntity;
  * @ORM\Entity(repositoryClass="Gesdinet\JWTRefreshTokenBundle\Entity\RefreshTokenRepository")
  * @UniqueEntity("refreshToken")
  */
-class RefreshToken implements RefreshTokenInterface
+class RefreshToken extends AbstractRefreshToken
 {
     /**
      * @var int
@@ -35,132 +35,10 @@ class RefreshToken implements RefreshTokenInterface
     protected $id;
 
     /**
-     * @var string
-     *
-     * @ORM\Column(name="refresh_token", type="string", length=128, unique=true)
-     * @Assert\NotBlank()
-     */
-    protected $refreshToken;
-
-    /**
-     * @var string
-     *
-     * @ORM\Column(name="username", type="string", length=255)
-     * @Assert\NotBlank()
-     */
-    protected $username;
-
-    /**
-     * @var string
-     *
-     * @ORM\Column(name="valid", type="datetime")
-     * @Assert\NotBlank()
-     */
-    protected $valid;
-
-    /**
-     * Get id.
-     *
-     * @return int
+     * {@inheritdoc}
      */
     public function getId()
     {
-        return $this->id;
-    }
-
-    /**
-     * Set refreshToken.
-     *
-     * @param string $refreshToken
-     *
-     * @return RefreshToken
-     */
-    public function setRefreshToken($refreshToken = null)
-    {
-        if (null == $refreshToken) {
-            $this->refreshToken = bin2hex(openssl_random_pseudo_bytes(64));
-        } else {
-            $this->refreshToken = $refreshToken;
-        }
-
-        return $this;
-    }
-
-    /**
-     * Get refreshToken.
-     *
-     * @return string
-     */
-    public function getRefreshToken()
-    {
-        return $this->refreshToken;
-    }
-
-    /**
-     * Set valid.
-     *
-     * @param \DateTime $valid
-     *
-     * @return RefreshToken
-     */
-    public function setValid($valid)
-    {
-        $this->valid = $valid;
-
-        return $this;
-    }
-
-    /**
-     * Get valid.
-     *
-     * @return \DateTime
-     */
-    public function getValid()
-    {
-        return $this->valid;
-    }
-
-    /**
-     * Set username.
-     *
-     * @param string $username
-     *
-     * @return RefreshToken
-     */
-    public function setUsername($username)
-    {
-        $this->username = $username;
-
-        return $this;
-    }
-
-    /**
-     * Get username.
-     *
-     * @return string
-     */
-    public function getUsername()
-    {
-        return $this->username;
-    }
-
-    /**
-     * Check if is a valid refresh token.
-     *
-     * @return bool
-     */
-    public function isValid()
-    {
-        $datetime = new \DateTime();
-
-        return ($this->valid >= $datetime) ? true : false;
-    }
-
-    /**
-     * @return string Refresh Token
-     */
-    public function __toString()
-    {
-        return $this->getRefreshToken();
+        $this->id;
     }
 }
